# config/swagger.yaml
openapi: 3.0.0
info:
  title: Khrestik API
  description:  Find out how your APIs work.
  version: 1.0.0
# info:
#     title: Khrestik API
#     description: Find out how your APIs work.
#     version: 1.0.0
servers:
    - url: http://localhost:8085/api/v1
    - url: http://95.179.185.226:8085/api/v1
paths:
    /product/all:
        get:
            description: "отримання усіх продуктів. задокументованими є 2 типи помилок: 500 - помилка сервера та 202 - дані успішно передано клієнту"
            summary: get all product data
            tags:
                - Products
            responses:
                '202':
                    description: all products received successfully
                    content:
                        application:json:
                            properties:
                                products:
                                    schema:
                                        $ref: '#/components/schemas/Product'
                                categories:
                                    schema:
                                        $ref: '#/components/schemas/Categories'
                '500':
                    description: server error
    /product/single/{productId}:
        get:
            description: "отримання одного конкретного продукту на основі переданого _id."
            summary: get single product data
            parameters:
                - in: path
                  name: productId
                  schema:
                    type: string
                  required: true
                  description: unique mongo id
            tags:
                - Products
            responses:
                '202':
                    description: all products received successfully
                    content:
                        application:json:
                            properties:
                                product:
                                    schema:
                                        $ref: '#/components/schemas/Product'
                                categories:
                                    schema:
                                        $ref: '#/components/schemas/Categories'
                '404':
                    description: id not found
                '422':
                    description: id probably incorrect
                '500':
                    description: server error 
    /product/store-form-data:
        post:
            description: "застосовується для збереження даних про окремі продукти. параметр images містить опис зображень продукту, та може містити до 50 зображень, переданих із вказанням індексу: images[0], images[1], ..., images[n(max 50)] виключно 2 параметри є обов'язковими: title та description . Продукт надсилається у вигляді FormData"
            tags:
                - Products
            requestBody:
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                isActive:
                                    type: boolean
                                    default: true
                                title:
                                    type: string
                                description:
                                    type: string
                                price:
                                    type: number
                                qty_available:
                                    type: integer
                                category:
                                    type: string
                                images[0][alt]:
                                    type: string
                                images[0][is_main]:
                                    type: boolean
                                    default: true
                                images[0]:
                                    type: file
                                images[1][alt]:
                                    type: string
                                images[1][is_main]:
                                    type: boolean
                                    default: true
                                images[1]:
                                    type: file
                            required:
                                - description
                                - title
            responses:
                '201':
                    description: created product
                    content:
                        application:json:
                            properties:
                                product:
                                    schema:
                                        $ref: '#/components/schemas/Product'
                                categories:
                                    schema:
                                        $ref: '#/components/schemas/Categories'
                '422':
                    description: incorrect data params passed
                '500':
                    description: server error 
    /product/edit/{product_id}:
        patch:
            description: "редагування уже існуючого продукту, а також додавання чи зміна даних супутніх зображень: alt, is_main. Для редагування іcнуючого зображення обов'язковим є передача параметра _id цього зображення. Інакше за відсутності параметру _id зображення буде розцінене як цілком нове і буде додано до масиву значень"
            parameters:
                -  in: path
                   name: product_id
                   schema:
                    type: string
                    required: true
            tags:
                - Products
            requestBody:
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                isActive:
                                    type: boolean
                                    default: true
                                title:
                                    type: string
                                description:
                                    type: string
                                price:
                                    type: number
                                qty_available:
                                    type: integer
                                category:
                                    type: string
                                images[0][_id]:
                                    type: string
                                images[0][alt]:
                                    type: string
                                images[0][is_main]:
                                    type: boolean
                                    default: true
                                images[0]:
                                    type: file
                                images[1][_id]:
                                    type: string
                                images[1][alt]:
                                    type: string
                                images[1][is_main]:
                                    type: boolean
                                    default: true
                                images[1]:
                                    type: file
                                
            responses:
                '202':
                    description: успішно відредаговано
                    content:
                        application:json:
                            properties:
                                product:
                                    schema:
                                        $ref: '#/components/schemas/Product'
                                categories:
                                    schema:
                                        $ref: '#/components/schemas/Categories'
                '404':
                    description: продукт не знайдено
                '422':
                    description: 'дані передано некоректно, або дані відсутні (за умови err: Cannot read property length of undefined")'
                '500':
                    description: server error 
    /product/image/{image_id}:
        delete:
            description: видалення зображення. Для видалення зображення потрібно передати _id зображення 
            parameters:
                - in: path
                  name: image_id
            tags:
                - Products
            responses:
                '200':
                    description: успішно видалено
                    content:
                        application:json:
                            properties:
                                product:
                                    schema:
                                        $ref: '#/components/schemas/Product'
                '404':
                    description: не знайдено
                                    
                '500':
                    description: server error
    /product/document/{doc_id}:
        delete:
            description: видалення цілого документу. Для видалення у запиті потрібно передати _id продукту. Каскадом при видаленні продукту видаляється оригінальне зображення
            parameters:
                - in: path
                  name: doc_id
            tags:
                - Products
            responses:
                '204':
                    description: успішно видалено
                    content:
                        application:json:
                            properties:
                                products:
                                    schema:
                                        $ref: '#/components/schemas/Product'
                '404':
                    description: не знайдено
                '500':
                    description: server error

components:
    schemas:
        Product:
            type: object
            properties:
                _id: 
                    type: string
                    required: true
                description:
                    type: string
                    required: true
                title:
                    type: string
                    required: true
                isActive:
                    type: boolean
                    default: true
                price: 
                    type: number
                qty_available:
                    type: integer
                createdAt:
                    type: string
                updatedAt:
                    type: string
                category:
                    type: string
                    required: true
                images: 
                    type: array
                    items:
                        type: object
                        properties:
                            alt:
                                type: string
                            url:
                                type: string
                            _id:
                                type: string
                            is_main:
                                type: boolean
        Image:
            type: object
            properties:
                alt: 
                    type: string
                is_main: 
                    type: boolean
                url:
                    type: string
                file: 
                    type: string
        Categories:
            type: array
            items:
                type: string